Adopt JSONFoundation for JSON value, schema, and JSON-RPC types (#146)
Replace SwiftMCP's vendored JSON-RPC and JSON Schema types — and the
re-exported JSONValue module — with one dependency on the renamed,
Foundation-only JSONFoundation package (1.1.0), which now bundles all three
layers under a single module.

- Delete the local JSONRPC (JSONRPCID, JSONRPCMessage) and JSON Schema
  (JSONSchema(+Codable/+Encoding), SchemaRepresentable, SchemaMetadata,
  SchemaPropertyInfo, JSONSchemaTypeConvertible, AnyCodingKey, caseLabels)
  sources; they now live in JSONFoundation.
- Exports.swift re-exports JSONFoundation via @_exported, so `import SwiftMCP`
  still surfaces these types — no per-file import churn.
- JSONRPCID's numeric case is `.integer` (was `.int`), matching JSONValue's
  vocabulary; updated all call sites.
- JSONRPCMessage+Batch keeps only the protocol-version batching gate; the pure
  decode helpers (decodeMessages/isBatchPayload) moved to JSONFoundation.
- Error payload `data` is now JSONValue?; LogMessage adopts the non-throwing
  JSONValue(jsonObject:).

All 476 tests pass; build is warning-free.

+++ Tests/SwiftMCPTests/MCPServerTests.swift
@@ -26,7 +26,7 @@ private func extractInitializeResponseResult(_ message: JSONRPCMessage) throws -
         throw TestError("Expected response case")
     }
     #expect(response.jsonrpc == "2.0")
-    #expect(response.id == .int(1))
+    #expect(response.id == .integer(1))
     guard let result = response.result else {
         throw TestError("Result is missing")
     }
@@ -104,7 +104,7 @@ func testToolsListRequest() async throws {
         return
     }
     #expect(response.jsonrpc == "2.0")
-    #expect(response.id == .int(2))
+    #expect(response.id == .integer(2))
     #expect(response.result != nil)
 
     guard let result = response.result else {
@@ -152,7 +152,7 @@ func testToolCallRequest() async throws {
         return
     }
     #expect(response.jsonrpc == "2.0")
-    #expect(response.id == .int(3))
+    #expect(response.id == .integer(3))
     #expect(response.result != nil)
 
     guard let result = response.result else {
@@ -193,7 +193,7 @@ func testToolCallRequestWithError() async throws {
         return
     }
     #expect(response.jsonrpc == "2.0")
-    #expect(response.id == .int(4))
+    #expect(response.id == .integer(4))
     #expect(response.result != nil)
 
     guard let result = response.result else {
@@ -244,7 +244,7 @@ func testToolCallRequestWithInvalidArgument() async throws {
         return
     }
     #expect(response.jsonrpc == "2.0")
-    #expect(response.id == .int(5))
+    #expect(response.id == .integer(5))
     #expect(response.result != nil)
 
     guard let result = response.result else {
@@ -274,7 +274,7 @@ func testCustomNameAndVersion() async throws {
     let calculator = CustomNameCalculator()
 
     // Get the response using our test method
-    let response = await calculator.createInitializeResponse(id: .int(1))
+    let response = await calculator.createInitializeResponse(id: .integer(1))
 
     // Extract server info from the response using pattern matching
     guard case .response(let responseData) = response else {
@@ -305,7 +305,7 @@ func testDefaultNameAndVersion() async throws {
     let calculator = DefaultNameCalculator()
 
     // Get the response using our test method
-    let response = await calculator.createInitializeResponse(id: .int(1))
+    let response = await calculator.createInitializeResponse(id: .integer(1))
 
     // Extract server info from the response using pattern matching
     guard case .response(let responseData) = response else {
@@ -351,7 +351,7 @@ func testUnknownMethodReturnsMethodNotFoundError() async throws {
         return
     }
     #expect(response.jsonrpc == "2.0")
-    #expect(response.id == .int(99))
+    #expect(response.id == .integer(99))
     #expect(response.error.code == -32601)
     #expect(response.error.message == "Method not found")
 }
